requirements.txt
streamlit
pandas
python-dotenv
PyPDF2
plotly
groq